task_1/requirements.txt
alembic==1.13.3
asyncpg==0.29.0
certifi==2024.8.30
charset-normalizer==3.4.0
greenlet==3.1.1
idna==3.10
Mako==1.3.5
MarkupSafe==3.0.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
